pyproject.toml
[tool.poetry]
name = "calliope"
version = "0.1.0"
description = ""
authors = ["David Meredith <daviddmeredith@gmail.com>"]
license = "ISC"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
scipy = "^1.14.1"
numpy = "^2.1.3"
pyaudio = "^0.2.14"
pygame = "^2.6.1"
pyopengl = "^3.1.7"
pyopengl-accelerate = "^3.1.7"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
